# docs/swagger.yaml
definitions:
  github_com_maidulcu_masaar-crm_internal_domain.Contact:
    properties:
      assigned_to:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      language:
        description: '"ar" | "en"'
        type: string
      lead_score:
        type: integer
      phone_wa:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_maidulcu_masaar-crm_internal_domain.MessageDirection:
    enum:
    - inbound
    - outbound
    type: string
    x-enum-varnames:
    - DirectionInbound
    - DirectionOutbound
...
  github_com_maidulcu_masaar-crm_internal_domain.ThreadStatus:
    enum:
    - open
    - closed
    - pending
    type: string
    x-enum-varnames:
    - ThreadOpen
    - ThreadClosed
    - ThreadPending
...
  github_com_maidulcu_masaar-crm_internal_domain.WhatsAppMessage:
    properties:
      body:
        type: string
      direction:
        $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.MessageDirection'
      id:
        type: string
      media_url:
        type: string
      sent_at:
        type: string
      thread_id:
        type: string
      wa_message_id:
        type: string
    type: object
  github_com_maidulcu_masaar-crm_internal_domain.WhatsAppThread:
    properties:
      ai_summary:
        type: string
      contact:
        allOf:
        - $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.Contact'
        description: Joined
      contact_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      last_message_at:
        type: string
      message_count:
        type: integer
      messages:
        items:
          $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.WhatsAppMessage'
        type: array
      thread_status:
        $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.ThreadStatus'
      wa_account_id:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /threads:
    get:
      description: Returns paginated WhatsApp conversation threads with contact details.
      parameters:
      - description: 'Filter by status: open|pending|closed'
        in: query
        name: status
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.WhatsAppThread'
            type: array
      security:
      - BearerAuth: []
      summary: List WhatsApp threads
      tags:
      - WhatsApp
  /threads/{id}/close:
    post:
      description: Sets thread_status to closed.
      parameters:
      - description: Thread UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Close thread
      tags:
      - WhatsApp
  /threads/{id}/messages:
    get:
      description: Returns messages for a WhatsApp thread, ordered by sent_at ascending.
      parameters:
      - description: Thread UUID
        in: path
        name: id
        required: true
        type: string
      - description: Max messages to return (default 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_maidulcu_masaar-crm_internal_domain.WhatsAppMessage'
            type: array
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get thread messages
      tags:
      - WhatsApp
swagger: "2.0"